Refactor inheritance Sql strategies https://github.com/SoftUni-Internal/exam-systems-issues/issues/545
